next-lane-self-drives/python-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.3.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
httpx==0.27.0
python-multipart==0.0.9
